refactor: moved back lib controller for posts and comments outside groups folder
